Fix review comments: move elements to correct package, fix imports
