+++ .github/workflows/ci.yml
@@ -2,7 +2,11 @@ name: Run Tests
 
 on:
   push:
+    branches:
+      - main
   pull_request:
+    branches:
+      - main
 
 jobs:
   test:
