+++ src/main/java/elements/Button.java
@@ -1,9 +1,10 @@
-package pages;
+package elements;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
+import pages.BaseElement;
 
 public class Button extends BaseElement {
