+++ src/main/java/elements/Label.java
@@ -1,7 +1,8 @@
-package pages;
+package elements;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+import pages.BaseElement;
 
 public class Label extends BaseElement {
