+++ src/main/java/elements/MenuItem.java
@@ -1,4 +1,4 @@
-package pages;
+package elements;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.NoSuchElementException;
